Temporary commit

+++ Source/LoadingExtension.cs
@@ -12,7 +12,7 @@ namespace DifficultyTuningMod
 
             if (mode == LoadMode.NewGame)
             {
-                long moneyToAdd = (d.InitialMoney.Value - 70) * 100000;
+                int moneyToAdd = (d.InitialMoney.Value - 70) * 100000;
                 Singleton<EconomyManager>.instance.AddResource(EconomyManager.Resource.LoanAmount, moneyToAdd, ItemClass.Service.None, ItemClass.SubService.None, ItemClass.Level.None);
             }
